Context Unification, Phase 3 follow-up
