+++ packages/make-meaning/src/graph-context.ts
@@ -23,19 +23,6 @@ import type { ResourceDescriptor } from '@semiont/core';
 // hand-written local twins were deleted — this is the one canonical type definition.
 type KnowledgeGraph = components['schemas']['KnowledgeGraph'];
 
-/**
- * Flattened neighborhood views derived from a `KnowledgeGraph` (CONTEXT-UNIFICATION P3, Q1=A).
- * One derivation, shared by the annotation producer's `inferredRelationshipSummary` and the
- * matcher (P4). It reports what the graph holds — and the graph is a projection of the event log
- * (the system of record), read here because it is the queryable projection at gather time.
- */
-export interface GraphViews {
-  connections: { resourceId: string; resourceName: string; entityTypes: string[]; bidirectional: boolean }[];
-  citedBy: { resourceId: string; resourceName: string }[];
-  citedByCount: number;
-  siblingEntityTypes: string[];
-}
-
 export class GraphContext {
   /**
    * Get all resources referencing this resource (backlinks)
@@ -125,9 +112,7 @@ export class GraphContext {
       const peerId = getResourceId(conn.targetResource);
       if (!peerId) continue;
       addResourceNode(peerId, conn.targetResource.name, getResourceEntityTypes(conn.targetResource));
-      if (mainId) {
-        edges.push({ source: mainId, target: peerId, type: conn.relationshipType || 'link', bidirectional: conn.bidirectional });
-      }
+      edges.push({ source: mainId, target: peerId, type: conn.relationshipType || 'link', bidirectional: conn.bidirectional });
     }
 
     // Inbound citations → citing-resource nodes + `citation` edges (citing → main)
@@ -138,9 +123,7 @@ export class GraphContext {
       citedSeen.add(source);
       const view = await kb.views.get(createResourceId(source));
       addResourceNode(source, view?.resource?.name ?? source, view?.resource ? getResourceEntityTypes(view.resource) : []);
-      if (mainId) {
-        edges.push({ source, target: mainId, type: 'citation' });
-      }
+      edges.push({ source, target: mainId, type: 'citation' });
     }
 
     // Annotations on this resource → annotation nodes + `annotation-of` edges
@@ -148,58 +131,9 @@ export class GraphContext {
       if (!ann.id || seen.has(ann.id)) continue;
       nodes.push({ id: ann.id, type: 'annotation', label: ann.motivation ?? 'annotation', entityTypes: getEntityTypes(ann) });
       seen.add(ann.id);
-      if (mainId) {
-        edges.push({ source: ann.id, target: mainId, type: 'annotation-of' });
-      }
+      edges.push({ source: ann.id, target: mainId, type: 'annotation-of' });
     }
 
     return { nodes, edges };
-  }
-
-  /**
-   * Derive the flattened views (connections / citedBy / siblings) from a `KnowledgeGraph`
-   * (CONTEXT-UNIFICATION P3, Q1=A). Reports the graph as-is — Option A: missing-view citers are
-   * kept (the citation edge reflects a real reference event), and the only filter is excluding the
-   * focal annotation from siblings (an annotation isn't its own sibling). Peer connections are
-   * edges out of `mainResourceId`; citations and `annotation-of` edges point INTO it.
-   */
-  static deriveViews(
-    graph: KnowledgeGraph,
-    mainResourceId: string,
-    focalAnnotationId?: string,
-  ): GraphViews {
-    const nodeById = new Map(graph.nodes.map((n) => [n.id, n] as const));
-
-    const connections: GraphViews['connections'] = [];
-    const citedBy: GraphViews['citedBy'] = [];
-    for (const edge of graph.edges) {
-      if (edge.type === 'citation') {
-        if (edge.target !== mainResourceId) continue;
-        const node = nodeById.get(edge.source);
-        citedBy.push({ resourceId: edge.source, resourceName: node?.label ?? edge.source });
-      } else if (edge.source === mainResourceId) {
-        const node = nodeById.get(edge.target);
-        connections.push({
-          resourceId: edge.target,
-          resourceName: node?.label ?? edge.target,
-          entityTypes: node?.entityTypes ?? [],
-          bidirectional: edge.bidirectional ?? false,
-        });
-      }
-    }
-
-    const siblingEntityTypes = new Set<string>();
-    for (const node of graph.nodes) {
-      if (node.type === 'annotation' && node.id !== focalAnnotationId) {
-        for (const et of node.entityTypes ?? []) siblingEntityTypes.add(et);
-      }
-    }
-
-    return {
-      connections,
-      citedBy,
-      citedByCount: citedBy.length,
-      siblingEntityTypes: Array.from(siblingEntityTypes),
-    };
   }
 }
